pyproject.toml
[project]
name = "nightjar"
version = "0.1.0"
description = "Nightjar — verified spec-as-source platform for AI-generated code"
requires-python = ">=3.11"
dependencies = [
    "click>=8.0",
    "litellm>=1.0",
    "pydantic>=2.0",
    "hypothesis>=6.0",
    "pyyaml>=6.0",
    "rich>=13.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
]

[project.scripts]
nightjar = "nightjar.cli:main"

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
markers = [
    "integration: tests requiring external services (LLM API, Dafny binary)",
]
